requirements.txt
pytesseract
openai
python-dotenv
kivy